frontend/api: move bitbox02 functions to bitbox02.ts
Easier structure, to have all bb02 api calls in one
file. E.g. the statusChanged subscription needs `TStatus`.
