+++ frontends/web/src/routes/device/bitbox02/wizard.tsx
@@ -17,8 +17,7 @@
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useLoad, useSync } from '@/hooks/api';
-import { getStatus, getVersion, verifyAttestation } from '@/api/bitbox02';
-import { attestationCheckDone, statusChanged } from '@/api/devicessync';
+import { attestationCheckDone, getStatus, getVersion, verifyAttestation, statusChanged } from '@/api/bitbox02';
 import { AppUpgradeRequired } from '@/components/appupgraderequired';
 import { FirmwareUpgradeRequired } from '@/routes/device/upgrade-firmware-required';
 import { Main } from '@/components/layout';
